made new controller LoadPostController
