+++ public/index.php
@@ -17,23 +17,10 @@ $router->add('delete_post', ['controller' => 'DeletePostController', 'action' =>
 $router->add('ajax_load', ['controller' => 'AjaxLoadPostController', 'action' => 'loadPostAjax']);
 $router->add('confirm_post', ['controller' => 'ConfirmPostController', 'action' => 'confirmPost']);
 $router->add('comment_frame', ['controller' => 'CommentController', 'action' => 'frameComment']);
-$router->add('post', ['controller' => 'PostsController', 'action' => 'loadPostsFriends']);
+$router->add('load_post', ['controller' => 'LoadPostController', 'action' => 'loadPost']);
 $router->add('submit_post', ['controller' => 'SubmitPostController', 'action' => 'submitPost']);
 $router->add('logged_out', ['controller' => 'LogoutController', 'action' => 'logout']);
 $router->add('{controller}/{action}');
 $router->add('{controller}/{id:\d+}/{action}');
 
 $router->dispatch($_SERVER['QUERY_STRING']);
-
-//echo 'Requested URL = "' . $_SERVER['QUERY_STRING'] . '"';
-
-/*
-$controller = match ($_SERVER['REQUEST_URI'])
-{
-    '/' => function() {
-        header("Location: files/homepage.php");
-    }
-    //'/addPost' => 
-};
-$controller();
-*/
